backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
http
gunicorn
whitenoise
